Remove default trait validation from loader
With this change the IDL loader won't throw a syntax error when you
try to assign a default value to a list, map, or union member using
the default value syntactic sugar. This validation is already taken
care of after the model is loaded, and provides a better error
message.