Add ability to find a child at path
Adds a findChildAtPath method to TreeCursor which takes in a list
of TreeType, and returns the first descendant for which the types
of the trees along the path from the current tree to that
descendant match the provided list.

This is useful for extracting a specific child out of the tree when
you know the structure.